Bump TocEntries.SortOrder from tinyint to smallint
Ebooks can have more than 255 TocEntries. (572 is the current highest.)

+++ config/sql/se/TocEntries.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `TocEntries` (
   `EbookId` int(10) unsigned NOT NULL,
   `TocEntry` text NOT NULL,
-  `SortOrder` tinyint(3) unsigned NOT NULL,
+  `SortOrder` smallint unsigned NOT NULL,
   KEY `index1` (`EbookId`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
